+++ packages/camera/camera_android_camerax/android/src/main/java/io/flutter/plugins/camerax/PreviewHostApiImpl.java
@@ -89,6 +89,9 @@ public class PreviewHostApiImpl implements PreviewHostApi {
               }
 
               @Override
+              // TODO(bparrishMines): Replace with onSurfaceCleanup once available on stable. See
+              // https://github.com/flutter/flutter/issues/161256.
+              @SuppressWarnings({"deprecation", "removal"})
               public void onSurfaceDestroyed() {
                 // Invalidate the SurfaceRequest so that CameraX knows to to make a new request
                 // for a surface.